projects/AcronymsDefined/requirements.txt
pdfplumber>=0.5.28
Pillow>=9.0.0

